refactor
re-use the steps
