+++ tests/e2e/support/types.ts
@@ -15,7 +15,7 @@ export interface Actor {
   context: BrowserContext
   page: Page
   close(): Promise<void>
-  updatePage(page: Page): Promise<Page>
+  newPage(page: Page): Promise<Page>
 }
 
 export interface User {
